Asya/requirements.txt
# --- Ядро ---
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
pydantic>=2.8.0
loguru>=0.7.0
PyYAML>=6.0
numpy>=1.24.0
pandas>=2.2.0
uuid>=1.30

# --- Аудио и обработка ---
torchaudio>=2.4.0
librosa>=0.10.0
soundfile>=0.12.0
pydub>=0.25.1
ffmpeg-python>=0.2.0

# --- NeMo (STT / VAD) ---
nemo_toolkit[asr]

# --- Silero TTS ---
silero==0.7.0

# --- llama.cpp (для GGUF-моделей) ---
llama-cpp-python>=0.3.0

# --- Вспомогательные ---
thefuzz>=0.19.0
python-Levenshtein>=0.21.0

# --- Опционально: SQLite (Python использует sqlite3 по умолчанию, но можно явно указать) ---
# Встроенная библиотека sqlite3 не нуждается в установке, но если нужна явная зависимость:
pysqlite3-binary>=0.5.0

# --- Для тестов (не обязателен в продакшене) ---
pytest>=8.0.0
pytest-asyncio>=0.23.0

# --- Дополнительно: мониторинг ---
psutil>=6.0.0
